Move design from home to discover screen
