+++ features/home/src/main/res/values-ro-rRO/strings.xml
@@ -4,7 +4,4 @@
     <string name="default_label">Teacher Workout</string>
     <string name="profile">Profile screen</string>
     <string name="home_landing">Home landing screen</string>
-    <string name="themes_title">Categorii de lecții</string>
-    <string name="discover_lessons_title">Descoperă lecțiile</string>
-    <string name="search_label">Caută un subiect</string>
 </resources>
